move also TmLiSession to session folder
